Adds an explrotary notebook for h2k parser
This commit adds a new exploratory notebook. The purpose of this
analysis is to build a h2k parser. Adam W. provided a code for parsing
h2k xml tree. The notebook adopts and adapt the algorithm to work with
the xmlschema. The work is still in progress. The code currently include
file info, weather data, house spec, and vintage. I will implement the
remainder of the Adam's code in next few days.

After completion of this step I will refactor the code and create a
class for h2k parsing.