Add the h2kparser to the src
This commit adds a module to the src directory. The module contains the
ParseH2K class which parse the building data, components, and equipment.
The class has multiple methods. Each method is created to report a
susbet of the h2k data. Majority of the data is reported as python
dictionaries.

A test directory is added to the branch that contains test module and
data. Pytest is used for creating the test. Additional tests should be
added for methods of the ParseH2K class.

ASHARAE140 models are added and will be used for test and validation.